requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
requests==2.32.3
pytest==8.3.3
rapidfuzz==3.9.0
